requirements.txt
flask
flask-cors
flask-pymongo
flask-jwt-extended
PyPDF2
python-dotenv

google-generativeai==0.8.0

langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.53
pydantic==1.10.12
